+++ components/input/index.js
@@ -2,7 +2,7 @@ import Input from './Input';
 import InputBase from './InputBase';
 import SingleLineInputBase from './SingleLineInputBase';
 import NumericInput from './NumericInput';
-import TextArea from './TextArea';
+import Textarea from './Textarea';
 
-export { InputBase, NumericInput, SingleLineInputBase, TextArea };
+export { InputBase, NumericInput, SingleLineInputBase, Textarea };
 export default Input;